organized all files and added peripherals and system_comm (for communication) files
