+++ includes/system/system.h
@@ -70,10 +70,6 @@ extern "C" {
     //#define BAUDRATE 57600
 #define BRGVAL   ((FCY/BAUDRATE)/16)-1
 
-    // Current ADC buffer dimension
-#define ADC_CHANNELS 2
-#define ADC_BUFF 64
-#define TOT_ADC_BUFF ADC_CHANNELS * ADC_BUFF
 
 #define NUM_PROCESS_DEFAULT 2
 #define PROCESS_IDLE 0
@@ -85,39 +81,6 @@ extern "C" {
         uint8_t priority;
         uint8_t frequency;
     } process_t;
-    
-    
-#ifdef UNAV_V1
-/// Number of available LEDs
-#define LED_NUM 4
-/// LED 1 - Green
-#define LED1_BIT _LATC6          // Led 1 Green
-#define LED1 0                   // Led 1 Green
-/// LED 2 - Red
-#define LED2_BIT _LATC7          // Led 2 Red
-#define LED2 1                   // Led 2 Red
-/// LED 3 - Yellow
-#define LED3_BIT _LATC8          // Led 3 Yellow
-#define LED3 2                   // Led 3 Yellow
-/// LED 4 - Blue
-#define LED4_BIT _LATC9          // Led 4 Blue
-#define LED4 3                   // Led 4 Blue
-#elif ROBOCONTROLLER_V3
-/// Number of available LEDs
-#define LED_NUM 2
-/// LED 1 - Green
-#define LED1_BIT _LATA8          // Led 1 green
-#define LED1 0                   // Led 1 green
-/// LED 2 - Green
-#define LED2_BIT _LATA9          // Led 2 green
-#define LED2 1                   // Led 2 green
-#elif MOTION_CONTROL
-/// Number of available LEDs
-#define LED_NUM 1
-/// LED 1 - Green
-#define LED1_BIT _LATA4          // Led Blue
-#define LED1 0                   // Led Blue
-#endif
 
     /******************************************************************************/
     /* System Function Prototypes                                                 */
@@ -132,21 +95,7 @@ extern "C" {
      */
     void ConfigureOscillator(void);
     
-    /**
-     * I/O and Peripheral Initialization
-     */
-    void Peripherals_Init(void);
-    
-    /**
-     * Initialization led blink
-     */
-    void InitLEDs(void);
-    /**
-     * Update frequency or type of blink
-     * @param led to control
-     * @param blink number of blinks
-     */
-    inline void UpdateBlink(short num, short blink);
+    inline uint16_t get_Frequency(void);
     
     /**
      * Initialization all system events
@@ -189,18 +138,6 @@ extern "C" {
     /** Initialization Timer 1 - Timer system
      */
     void InitTimer1(void);
-
-    /** Initialization Timer 2 for IC (Input Capture)
-     */
-    void InitTimer2(void);
-
-    /** Initialization DMA0 for ADC current
-     */
-    void InitDMA0(void);
-
-    /** Initialization ADC for measure current motors
-     */
-    void InitADC(void);
 
 #ifdef	__cplusplus
 }
