+++ src/configuration_bits.c
