+++ src/main.c
@@ -33,9 +33,12 @@
 #include <stdint.h>        /* Includes uint16_t definition                    */
 #include <stdbool.h>       /* Includes true/false definition                  */
 
-#include "system/system.h" /* System funct/params, like osc/peripheral config */
-
 #include <system/task_manager.h>
+
+#include "system/system.h" /* System funct/params, like osc/peripheral config */
+#include "system/system_comm.h"
+#include "system/peripherals.h"
+
 
 #include "communication/serial.h"
