+++ src/motors/motor_init.c
@@ -177,6 +177,23 @@ void InitIC(short motIdx) {
     }
 }
 
+void InitTimer2(void) {
+    //T2CON = 10100000 00000000
+    T2CONbits.TON = 0; // Disable Timer
+    T2CONbits.TSIDL = 1; // Stop in Idle Mode bit
+    T2CONbits.TGATE = 0; // Disable Gated Timer mode
+    T2CONbits.TCKPS = 0b00; // Select 1:1 Prescaler
+    T2CONbits.TCS = 0; // Select internal clock source
+    TMR2 = 0x00; // Clear timer register
+    PR2 = TMR2_VALUE; // Load the period value
+
+    IPC1bits.T2IP = PWM_TIMER_LEVEL; // Set Timer 1 Interrupt Priority Level
+    IFS0bits.T2IF = 0; // Clear Timer 1 Interrupt Flag
+    IEC0bits.T2IE = 1; // Enable Timer1 interrupt
+
+    T2CONbits.TON = 1; // Start Timer
+}
+
 void Motor_Init(short motIdx) {
     /* Open QEI */
     InitQEI(motIdx);
